requirements.txt
Flask==2.3.3
langchain==0.1.0
langchain-community==0.0.20
langchain-google-genai==0.0.6
google-generativeai==0.3.2
faiss-cpu==1.7.4
PyMuPDF==1.23.8
python-docx==0.8.11
docx2txt==0.8
unstructured==0.10.30
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
